requirements.txt
Django==5.0.2
djangorestframework==3.14.0
pillow==10.2.0
django-simple-history==3.5.0
django-axes==6.3.0
django-ipware==6.0.4